+++ SlackNet/WebApi/Responses/OauthV2AccessResponse.cs
@@ -1,34 +1,28 @@
-﻿namespace SlackNet.WebApi
+﻿using SlackNet;
+
+namespace SlackNet.WebApi
 {
-    public class OauthV2AccessResponse
+    public class OauthV2AccessResponse : OauthAccessResponse
     {
-        public string AccessToken { get; set; }
-        #nullable enable
+#nullable enable
+        /// <summary>
+        /// Only available for OAuth flows with Token Rotation enabled. See more: https://api.slack.com/authentication/rotation
+        /// </summary>
         public string? RefreshToken { get; set; }
+        /// <summary>
+        /// Only available for OAuth flows with Token Rotation enabled. See more: https://api.slack.com/authentication/rotation
+        /// </summary>
         public int? ExpiresIn { get; set; }
-        #nullable disable
-        public string Scope { get; set; }
 
-        public Space Team { get; set; }
-        public Space Enterprise { get; set; }
+        /// <summary>
+        /// Only available for with the identity.team scope.
+        /// </summary>
+        public WorkSpace? Team { get; set; }
+        /// <summary>
+        /// Only available for enterprise price tier.
+        /// </summary>
+        public WorkSpace? Enterprise { get; set; }
+#nullable disable
         public AuthedUser AuthedUser { get; set; }
-    }
-
-    public class Space
-    {
-        public string Id { get; set; }
-        public string Name { get; set; }
-    }
-
-    public class AuthedUser
-    {
-        public string Id { get; set; }
-        public string Scope { get; set; }
-        public string AccessToken { get; set; }
-        #nullable enable
-        public int? ExpiresIn { get; set; }
-        public string? RefreshToken { get; set; }
-        #nullable disable
-        public string TokenType { get; set; }
     }
 }
